fix(vercel): move auxiliary api modules outside serverless api dir
